# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Events API - Luma-Inspired Event Management",
        "description": "Open-source headless Laravel event management API from Maak. Features organizations, events, ticket types, registrations, QR check-ins, waitlist management, and team member invitations.",
        "contact": {
            "name": "Maak",
            "url": "https://github.com/maakwizardry/events"
        },
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "http://events.test/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "PHP Built-in Server"
        }
    ],
    "paths": {
...
        "/registrations": {
            "get": {
                "tags": [
                    "Registrations"
                ],
                "summary": "List my registrations",
                "description": "Get all registrations for the authenticated user",
                "operationId": "8af1a2eedf460f934b7a19799b110195",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user registrations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "uuid": {
                                                        "type": "string",
                                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "confirmed"
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "total_price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 50
                                                    },
                                                    "is_checked_in": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "registered_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/{event}/register": {
            "post": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Register for event (authenticated)",
                "description": "Register for an event as an authenticated user. User information is taken from the authenticated account.",
                "operationId": "c9e71fb02705d594b495433b2901736b",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ticket_type_id"
                                ],
                                "properties": {
                                    "ticket_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "custom_fields": {
                                        "type": "object",
                                        "example": {
                                            "dietary_restrictions": "vegetarian",
                                            "company": "Acme Corp"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "uuid": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "confirmed"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "total_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "qr_code_data": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error or registration not open"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/registrations/{registration}": {
            "get": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Get my registration details",
                "description": "Get details of a specific registration. Users can only view their own registrations.",
                "operationId": "9364e1da3619e9f6ecbbc763f8d87782",
                "parameters": [
                    {
                        "name": "registration",
                        "in": "path",
                        "description": "Registration UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - not your registration"
                    },
                    "404": {
                        "description": "Registration not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/registrations/{registration}/cancel": {
            "post": {
                "tags": [
                    "Registrations"
                ],
                "summary": "Cancel my registration",
                "description": "Cancel a registration. Users can only cancel their own registrations. Frees up capacity for others.",
                "operationId": "8893bdac551873958b6c0d863be5ed53",
                "parameters": [
                    {
                        "name": "registration",
                        "in": "path",
                        "description": "Registration UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registration cancelled successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized - not your registration"
                    },
                    "422": {
                        "description": "Registration is already cancelled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use the token from POST /auth/login or POST /auth/register",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}